requirements.txt
streamlit
selenium
pandas
python-dotenv
browser-use
